+++ 1305015_read_input.h
@@ -10,11 +10,14 @@
 #include <iostream>
 #include <vector>
 #include <sstream>
+#include <cmath>
+#include <limits>
+#include <fstream>
 
 using namespace std;
 
 vector<string> split(const string &s, char delim);
-void print_vector(vector<vector<int> > &graph);
-int read_input(string filename, vector<vector<int> > &graph);
+void print_vector(vector<vector<float> > &graph);
+int read_input(string filename, vector<vector<float> > &graph);
 
 #endif //AE_TSP_READ_INPUT_H
